pyproject.toml
[project]
name = "openfr"
version = "0.1.0"
description = "Financial Research Agent based on AKShare"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "OpenFR Team" }
]
keywords = ["finance", "agent", "akshare", "research", "llm"]

dependencies = [
    "akshare>=1.14.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "openai>=1.50.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
openfr = "openfr.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openfr"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: mark test as integration test",
    "slow: mark test as slow running",
]
